+++ infra/resources/ced/prod/locals.tf
@@ -53,13 +53,22 @@ locals {
 
       CED_PORTAL_FE_BASE_URL = "https://${module.portal_fe_static_web_app.custom_domain}"
       CED_PRODUCT_ID         = "prod-ced"
-      ADMIN_FISCAL_CODES     = ""
+      ADMIN_FISCAL_CODES     = "516984510c575da00a39bcfcbc7e31ca4295384940dad4d2fd39f6e402f660b4"
     }
 
     startup_probe_path   = "/api/info/startup"
     readiness_probe_path = "/api/info/readiness"
   }
 
+  browser_be_container_app_name = provider::dx::resource_name({
+    prefix          = local.prefix
+    environment     = local.env_short
+    location        = local.location
+    domain          = local.domain
+    name            = "browser-be"
+    instance_number = 1
+    resource_type   = "container_app"
+  })
 
   # Browser BE Container App configuration
   browser_be = {
@@ -68,11 +77,17 @@ locals {
     image = "ghcr.io/pagopa/io-growth/ced-browser-be:latest"
 
     app_settings = {
+      APPLICATIONINSIGHTS_ENTRA_ID_AUTH_ENABLED = "false"
+      APPINSIGHTS_SAMPLING_PERCENTAGE           = "100"
+      TELEMETRY_SERVICE_NAME                    = local.browser_be_container_app_name
+
       PORT            = "8080"
+
       POSTGRES_HOST   = "${module.postgresql.postgres.name}.postgres.database.azure.com"
       POSTGRES_PORT   = "6432"
       POSTGRES_DB     = azurerm_postgresql_flexible_server_database.ced_test.name
       POSTGRES_SSL    = "true"
+
       REDIS_ENDPOINT  = module.redis_dx.endpoint
       REDIS_TLS       = "true"
       AZURE_CLIENT_ID = module.common_container_app_environment.user_assigned_identity.client_id
@@ -80,14 +95,24 @@ locals {
       # FIMS SSO settings
       BASE_URL             = "https://browser.ced.pagopa.it"
       FIMS_AUDIT_CONTAINER = "ced-browser-logs"
-      FIMS_REDIRECT_URL    = "${module.ced_apim.gateway_url}/browser/fcb"
-      FIMS_SCOPE           = "openid profile fiscal_code"
+      FIMS_REDIRECT_URL    = "https://api.ced.pagopa.it/api/ced-browser/v1/fcb"
+      FIMS_SCOPE           = "openid profile lollipop"
       TEST_USERS           = "6960f673e4bf8cc073a32b3b4579bfdb97b50b8df29964bdea6fcd1576d16f82"
     }
 
     startup_probe_path   = "/api/info/startup"
     readiness_probe_path = "/api/info/readiness"
   }
+
+  card_request_be_container_app_name = provider::dx::resource_name({
+    prefix          = local.prefix
+    environment     = local.env_short
+    location        = local.location
+    domain          = local.domain
+    name            = "card-request-be"
+    instance_number = 1
+    resource_type   = "container_app"
+  })
 
   # Card Request BE Container App configuration
   card_request_be = {
@@ -96,7 +121,12 @@ locals {
     image = "ghcr.io/pagopa/io-growth/ced-card-request-be:latest"
 
     app_settings = {
+      APPLICATIONINSIGHTS_ENTRA_ID_AUTH_ENABLED = "false"
+      APPINSIGHTS_SAMPLING_PERCENTAGE           = "100"
+      TELEMETRY_SERVICE_NAME                    = local.card_request_be_container_app_name
+
       PORT            = "8080"
+
       REDIS_ENDPOINT  = module.redis_dx.endpoint
       REDIS_TLS       = "true"
       AZURE_CLIENT_ID = module.common_container_app_environment.user_assigned_identity.client_id
@@ -107,8 +137,8 @@ locals {
       # FIMS SSO settings
       BASE_URL             = "https://card.ced.pagopa.it"
       FIMS_AUDIT_CONTAINER = "ced-card-request-logs"
-      FIMS_REDIRECT_URL    = "${module.ced_apim.gateway_url}/card-request/fcb"
-      FIMS_SCOPE           = "openid profile fiscal_code"
+      FIMS_REDIRECT_URL    = "https://api.ced.pagopa.it/api/ced-card/v1/fcb"
+      FIMS_SCOPE           = "openid profile lollipop"
       TEST_USERS           = "6960f673e4bf8cc073a32b3b4579bfdb97b50b8df29964bdea6fcd1576d16f82"
     }